backend/requirements.txt
Django==4.2.5
gunicorn==21.2.0
psycopg2-binary==2.9.7
djangorestframework==3.14.0